requirements.txt
pdfplumber
openai==0.28
python-dotenv
requests
